Add migration script and shift response deletion to appropriate service
