+++ backend/src/services/review.ts
@@ -4,7 +4,6 @@ import authorisation from '../connectors/authorisation/index.js'
 import { AccessRequestDoc } from '../models/AccessRequest.js'
 import { CollaboratorEntry, ModelDoc, ModelInterface } from '../models/Model.js'
 import { ReleaseDoc } from '../models/Release.js'
-import { ResponseDoc } from '../models/Response.js'
 import Review, { ReviewDoc, ReviewInterface } from '../models/Review.js'
 import ReviewRoleModel, { ReviewRoleInterface } from '../models/ReviewRole.js'
 import { UserInterface } from '../models/User.js'
@@ -13,7 +12,6 @@ import { BadReq, Forbidden, InternalError, NotFound } from '../utils/error.js'
 import { handleDuplicateKeys } from '../utils/mongo.js'
 import log from './log.js'
 import { getModelById } from './model.js'
-import { getResponsesByParentIds } from './response.js'
 import { requestReviewForAccessRequest, requestReviewForRelease } from './smtp/smtp.js'
 
 // This should be replaced by using the dynamic schema
@@ -132,24 +130,6 @@ export async function removeReleaseReviews(user: UserInterface, modelId: string,
       throw InternalError('The requested release review could not be deleted.', {
         modelId,
         semver,
-        error,
-      })
-    }
-  }
-
-  const responses = await getResponsesByParentIds(
-    user,
-    reviews.flatMap((r) => r.id),
-  )
-  const responseDeletions: ResponseDoc[] = []
-  for (const response of responses) {
-    try {
-      responseDeletions.push(await response.delete())
-    } catch (error) {
-      throw InternalError('The requested response could not be deleted.', {
-        modelId,
-        semver,
-        responseId: response.id,
         error,
       })
     }
